UI: improve destination tx status tracking (#241)
previously, if the graph was down, L1 -> L2 transfers were not found.
this is now addressed by query filtering for `TransferFromL1Completed` events.

previously, L2 -> L1 transfers were not found. this is now addressed by
calling `destL1Bridge.isTransferIdSpent`.

previously, if the destination network was the same as the source network (during convert to hToken), the dest tx would be displayed indefinitely loading. this is now addressed by not displaying a symbol for the dest tx status.

* don't attempt to check dest status for approve txs

* check for matching recipient

* add test.todo for inputs beginning with a decimal

* display destination tx status in TxStatusModal

* do not show dest tx status if same network

* get and display # of tx confirmations

* add styles